refactor: introduce `xml::from_scalar!`
