+++ src/xml.rs
@@ -27,3 +27,36 @@ pub(crate) fn deserialize<T: FromXmlOwned>(xml: &str) -> Result<T, Error> {
 pub(crate) struct Epp<T> {
     pub(crate) data: T,
 }
+
+macro_rules! from_scalar {
+    ($name:ty, $scalar:tt) => {
+        impl<'xml> ::instant_xml::FromXml<'xml> for $name {
+            fn matches(id: ::instant_xml::Id<'_>, field: Option<::instant_xml::Id<'_>>) -> bool {
+                match field {
+                    Some(field) => id == field,
+                    None => false,
+                }
+            }
+
+            fn deserialize<'cx>(
+                into: &mut Self::Accumulator,
+                field: &'static str,
+                deserializer: &mut ::instant_xml::Deserializer<'cx, 'xml>,
+            ) -> Result<(), ::instant_xml::Error> {
+                let mut value = None;
+
+                $scalar::deserialize(&mut value, field, deserializer)?;
+
+                if let Some(value) = value {
+                    *into = Some(Self::from(value));
+                }
+
+                Ok(())
+            }
+
+            type Accumulator = Option<Self>;
+            const KIND: ::instant_xml::Kind = ::instant_xml::Kind::Scalar;
+        }
+    };
+}
+pub(crate) use from_scalar;
